+++ benchmarks/asyncband/waitgroup/wait.rs
@@ -16,7 +16,6 @@
 // under the License.
 
 use std::future::IntoFuture;
-use std::pin::pin;
 
 use asyncband::waitgroup::WaitGroup;
 use divan::Bencher;
@@ -26,84 +25,9 @@ use crate::support::bench_context;
 use crate::support::poll_pending;
 use crate::support::poll_pinned_ready;
 
-const WORKER_COUNTS: &[usize] = &[1, 8, 32];
-const THREAD_COUNTS: &[usize] = &[1, 2, 8, 32];
-const CONTENDED_SAMPLE_SIZE: u32 = 256;
+const WAITER_COUNTS: &[usize] = &[1, 8, 32];
 
-#[divan::bench]
-fn ready_empty(bencher: Bencher) {
-    let mut context = bench_context();
-
-    bencher.bench_local(|| {
-        let mut wait = pin!(WaitGroup::new().into_future());
-        poll_pinned_ready(wait.as_mut(), &mut context);
-        black_box(())
-    });
-}
-
-#[divan::bench]
-fn worker_round_trip(bencher: Bencher) {
-    let root = WaitGroup::new();
-
-    bencher.bench_local(|| black_box(root.clone()));
-}
-
-#[divan::bench(threads = THREAD_COUNTS, sample_size = CONTENDED_SAMPLE_SIZE)]
-fn worker_round_trip_contended(bencher: Bencher) {
-    let root = WaitGroup::new();
-
-    bencher.bench(|| black_box(root.clone()));
-}
-
-#[divan::bench]
-fn cancel_pending(bencher: Bencher) {
-    let mut context = bench_context();
-
-    bencher.bench_local(|| {
-        let root = WaitGroup::new();
-        let worker = root.clone();
-        {
-            let mut wait = pin!(root.into_future());
-            poll_pending(wait.as_mut(), &mut context);
-        }
-        drop(worker);
-        black_box(())
-    });
-}
-
-#[divan::bench]
-fn complete_waiter(bencher: Bencher) {
-    let mut context = bench_context();
-
-    bencher.bench_local(|| {
-        let root = WaitGroup::new();
-        let worker = root.clone();
-        let mut wait = pin!(root.into_future());
-        poll_pending(wait.as_mut(), &mut context);
-
-        drop(worker);
-        poll_pinned_ready(wait.as_mut(), &mut context);
-        black_box(())
-    });
-}
-
-#[divan::bench(args = WORKER_COUNTS)]
-fn complete_worker_batch(bencher: Bencher, worker_count: usize) {
-    let mut context = bench_context();
-
-    bencher.bench_local(|| {
-        let root = WaitGroup::new();
-        let workers = (0..worker_count).map(|_| root.clone()).collect::<Vec<_>>();
-        let mut wait = pin!(root.into_future());
-        poll_pending(wait.as_mut(), &mut context);
-
-        drop(workers);
-        poll_pinned_ready(wait.as_mut(), &mut context);
-        black_box(())
-    });
-}
-
-#[divan::bench(args = WORKER_COUNTS)]
+#[divan::bench(args = WAITER_COUNTS)]
 fn complete_waiter_batch(bencher: Bencher, waiter_count: usize) {
     let mut context = bench_context();